% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cal_LZZL20}
\alias{cal_LZZL20}
\title{Bootstrapped LZZL statistics}
\usage{
cal_LZZL20(X, e, n_boundary, s0, p, sig)
}
\arguments{
\item{X}{\eqn{n\times p}}

\item{e}{multipliers, \eqn{n\times B}}

\item{n_boundary}{\code{[n_boundary:(n - n_boundary)]}}

\item{s0}{\eqn{s_0 = p / 2}}

\item{p}{\eqn{p \in \{1,2,3,4,5,\infty\}}}

\item{sig}{estimates of component-wise standard deviations}
}
\value{
Bootstrapped LZZL statistics
}
\description{
Bootstrapped LZZL statistics
}
\references{
Liu, B., Zhou, C., Zhang, X. and Liu, Y. (2020). A unified data-adaptive framework for high dimensional change point detection. Journal of the Royal Statistical Society: Series B (Statistical Methodology) 82 933–63.
}
\seealso{
\code{\link{YC21}} \code{\link{norm_s0_p}}
}
