% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cal_WZWY19}
\alias{cal_WZWY19}
\title{Normalization of Sum(0.5) statistic}
\usage{
cal_WZWY19(X)
}
\arguments{
\item{X}{\eqn{n\times p}}
}
\value{
Estimated expectation and variance of the Sum(0.5) statistic
}
\description{
Normalization of Sum(0.5) statistic
}
\references{
Wang, Y., Zou, C., Wang, Z. and Yin, G. (2019) Multiple change-points detection in high dimension. Random Matrices: Theory and Applications, 08, 1950014.
}
\seealso{
\code{\link{WZWY19}}
}
