% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cal_YC21}
\alias{cal_YC21}
\title{Bootstrapped Max(0.5) statistics}
\usage{
cal_YC21(X, e, n_boundary)
}
\arguments{
\item{X}{\eqn{n\times p}}

\item{e}{multipliers, \eqn{n\times B}}

\item{n_boundary}{\code{[n_boundary:(n - n_boundary)]}}
}
\value{
Bootstrapped Max(0.5) statistics
}
\description{
Bootstrapped Max(0.5) statistics
}
\references{
Yu, M. and Chen, X. (2021) Finite Sample Change Point Inference and Identification for High-Dimensional Mean Vectors. Journal of the Royal Statistical Society Series B: Statistical Methodology, 83, 247–270
}
\seealso{
\code{\link{YC21}}
}
