% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cal_ZWS22}
\alias{cal_ZWS22}
\title{ZWS statistics}
\usage{
cal_ZWS22(X, q, truc)
}
\arguments{
\item{X}{\eqn{n\times p}}

\item{q}{\eqn{q \in \{2, 6\}}, suggested}

\item{truc}{\eqn{\code{truc} = q}}
}
\value{
ZWS statistics
}
\description{
The following implementation is an adaptation of the Matlab code created by Zhang, Wang, and Shao (2020).
The original Matlab code can be accessed at the following link:
https://doi.org/10.1080/01621459.2021.1884562.
}
\references{
Zhang, Y., Wang, R. and Shao, X. (2022) Adaptive Inference for Change Points in High-Dimensional Data. Journal of the American Statistical Association, 117, 1751–1762.
}
