% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cusum}
\alias{cusum}
\title{CUSUM statistics}
\usage{
cusum(X, gamma)
}
\arguments{
\item{X}{\eqn{n\times p}, component-wise unit variances}

\item{gamma}{0 = CUSUM, 0.5 = two-sample}
}
\value{
CUSUM statistics, \eqn{(n - 1)\times p\times \code{length(gamma)}}
}
\description{
CUSUM statistics
}
