% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cusum_2sample_boots}
\alias{cusum_2sample_boots}
\title{Bootstrapped CUSUM statistics (two-sample version)}
\usage{
cusum_2sample_boots(X, e)
}
\arguments{
\item{X}{\eqn{n}-dimensional, unit variance}

\item{e}{multipliers, \eqn{n\times B}}
}
\value{
Bootstrapped CUSUM statistics, \eqn{(n - 1)\times B}
}
\description{
Bootstrapped CUSUM statistics (two-sample version)
}
