% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMS.R
\name{DMS}
\alias{DMS}
\title{DMS method for adaptive changepoint testing (for independent observations)}
\usage{
DMS(X, gam = 0, lam = NULL, type = "Fisher")
}
\arguments{
\item{X}{A data matrix of dimension \eqn{n\times p}, where each row represents an observation, and each column stands for a variable.}

\item{gam}{A numeric value (\eqn{0} or \eqn{0.5}) indicating which version of the DMS statistic to use.}

\item{lam}{An integer between \eqn{1} and \eqn{n/2}, representing a pre-specified boundary removal parameter if \code{gam = 0.5}.}

\item{type}{A string indicating the method used to combine p-values from the max- and sum-based tests. Accepted values are "min", "Fisher", or "Cauchy".}
}
\value{
\code{DMS} returns a list containing p-values of the DMS, max-, and sum-based testing procedures, respectively.
}
\description{
DMS method for adaptive changepoint testing (for independent observations)
}
\examples{
library(DMS)
set.seed(6)
n <- 200
p <- 200
tau <- floor(n / 2)
mu <- matrix(0, n, p)
mu[(tau + 1):n, 1:10] <- 0.5
eps <- matrix(rnorm(n * p), n, p)
X <- mu + eps
DMS(X, gam = 0)
# $pv.DMS
# [1] 1.528718e-07
# $pv.Max
# [1] 0.0004963364
# $pv.Sum
# [1] 1.565607e-05
DMS(X, gam = 0.5, lam = 20)
# $pv.DMS
# [1] 1.404542e-06
# $pv.Max
# [1] 0.005177261
# $pv.Sum
# [1] 1.565607e-05
}
