% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{est_var_diff}
\alias{est_var_diff}
\title{Difference-based variance estimates}
\usage{
est_var_diff(X)
}
\arguments{
\item{X}{\eqn{n\times p}}
}
\value{
Component-wise variance estimates (\eqn{1\times p})
}
\description{
Difference-based variance estimates
}
\references{
Rice, J. (1984). Bandwidth choice for nonparametric regression. The Annals of Statistics, 12, 1215-1230.
}
