% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{J15}
\alias{J15}
\title{Max(0) (Jirak, 2015)}
\usage{
J15(X, CUSUM = NULL, mod = TRUE)
}
\arguments{
\item{X}{\eqn{n\times p}, component-wise unit variances}

\item{CUSUM}{\eqn{(n - 1)\times p}, or \code{NULL}}

\item{mod}{\code{TRUE} = use modified normalizing factors in Wang and Feng (2023)}
}
\value{
The p-value and the test statistic
}
\description{
Max(0) (Jirak, 2015)
}
\references{
Jirak, M. (2015) Uniform change point tests in high dimension. The Annals of Statistics, 43, 2451–2483.

Wang, G. and Feng, L. (2023) Computationally efficient and data-adaptive changepoint inference in high dimension. Journal of the Royal Statistical Society Series B: Statistical Methodology, 85, 936–958.
}
