% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{LZZL20}
\alias{LZZL20}
\title{LZZL20}
\usage{
LZZL20(
  X,
  CUSUM = NULL,
  CUSUM_2sample = NULL,
  LZZL.s0 = ncol(X)/2,
  LZZL.p = c(1:5, Inf),
  LZZL.B = 500,
  LZZL.boundary = 0.2 * nrow(X),
  est = FALSE
)
}
\arguments{
\item{X}{\eqn{n\times p}}

\item{CUSUM}{CUSUM(0)}

\item{CUSUM_2sample}{CUSUM(0.5), for \eqn{\hat{\sigma}_{t,j}}}

\item{LZZL.s0}{\eqn{s_0 = p / 2}}

\item{LZZL.p}{\eqn{p \in \{1,2,3,4,5,\infty\}}}

\item{LZZL.B}{number of bootstrap replications}

\item{LZZL.boundary}{\code{[n.boundary:(n - n.boundary)]}}

\item{est}{\code{TRUE} = enable changepoint estimation}
}
\value{
The p-value
}
\description{
LZZL20
}
\references{
Liu, B., Zhou, C., Zhang, X. and Liu, Y. (2020) A Unified Data-Adaptive Framework for High Dimensional Change Point Detection. Journal of the Royal Statistical Society Series B: Statistical Methodology, 82, 933–963.
}
\seealso{
\code{\link{norm_s0_p}} \code{\link{cal_LZZL20}}
}
