% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{max_2sample}
\alias{max_2sample}
\title{Max(0.5) (Wang and Feng, 2023)}
\usage{
max_2sample(X, n.boundary, CUSUM = NULL)
}
\arguments{
\item{X}{\eqn{n\times p}, component-wise unit variances}

\item{n.boundary}{\code{[n.boundary:(n - n.boundary)]}}

\item{CUSUM}{\eqn{(n - 1)\times p}, or \code{NULL}}
}
\value{
The p-value and the test statistic
}
\description{
Max(0.5) (Wang and Feng, 2023)
}
\references{
Wang, G. and Feng, L. (2023) Computationally efficient and data-adaptive changepoint inference in high dimension. Journal of the Royal Statistical Society Series B: Statistical Methodology, 85, 936–958.
}
