% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{norm_s0_p}
\alias{norm_s0_p}
\title{\eqn{(s_0, p)}-norm}
\usage{
norm_s0_p(C, n_boundary, s0, p)
}
\arguments{
\item{C}{\eqn{(n - 1)\times p}}

\item{n_boundary}{\code{[n_boundary:(n - n_boundary)]}}

\item{s0}{\eqn{s_0}}

\item{p}{\eqn{p}}
}
\value{
Maximum and maximizer over all feasible changepoints for each \eqn{p}
}
\description{
\eqn{(s_0, p)}-norm
}
\references{
Liu, B., Zhou, C., Zhang, X. and Liu, Y. (2020). A unified data-adaptive framework for high dimensional change point detection. Journal of the Royal Statistical Society: Series B (Statistical Methodology) 82 933–63.
}
\seealso{
\code{\link{YC21}} \code{\link{cal_LZZL20}}
}
