% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{pgumbel}
\alias{pgumbel}
\title{Distribution function of the Gumbel distribution}
\usage{
pgumbel(x, mu = 0, beta = 1)
}
\arguments{
\item{x}{vector of quantiles}

\item{mu}{location}

\item{beta}{scale}
}
\description{
Distribution function of the Gumbel distribution
}
