% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{pvc}
\alias{pvc}
\title{P-value combination}
\usage{
pvc(pvs, type = "min")
}
\arguments{
\item{pvs}{p-values}

\item{type}{combination rule}
}
\value{
The combined p-value
}
\description{
P-value combination
}
