% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{WZWY19}
\alias{WZWY19}
\title{Sum(0.5), with power-enhanced version (Wang, Zou, Wang and Yin, 2019)}
\usage{
WZWY19(
  X,
  CUSUM = NULL,
  PE = TRUE,
  PE.c = 100,
  PE.h = (2 * log(nrow(X) * ncol(X)))^1.1,
  n.boundary = 0.1 * nrow(X)
)
}
\arguments{
\item{X}{\eqn{n\times p}}

\item{CUSUM}{\eqn{(n - 1)\times p}, or \code{NULL}}

\item{PE}{\code{TRUE} = include the power-enhanced version}

\item{PE.c}{\eqn{c}}

\item{PE.h}{\eqn{h}}

\item{n.boundary}{\code{[n.boundary:(n - n.boundary)]} (for Max(0.5))}
}
\value{
The p-value and the test statistic
}
\description{
Sum(0.5), with power-enhanced version (Wang, Zou, Wang and Yin, 2019)
}
\references{
Wang, Y., Zou, C., Wang, Z. and Yin, G. (2019) Multiple change-points detection in high dimension. Random Matrices: Theory and Applications, 08, 1950014.
}
\seealso{
\code{\link{cal_WZWY19}}
}
