% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{YC21}
\alias{YC21}
\title{Max(0.5), bootstrap calibrated (Yu and Chen, 2021)}
\usage{
YC21(X, n.boundary = 40, B = 200, CUSUM = NULL, M = NULL)
}
\arguments{
\item{X}{\eqn{n\times p}}

\item{n.boundary}{\code{[n.boundary:(n - n.boundary)]}}

\item{B}{number of bootstrap replications}

\item{CUSUM}{\eqn{(n - 1)\times p}, or \code{NULL}}

\item{M}{Max(0.5) statistic, or \code{NULL}}
}
\value{
The p-value and the test statistic
}
\description{
Max(0.5), bootstrap calibrated (Yu and Chen, 2021)
}
\references{
Yu, M. and Chen, X. (2021) Finite Sample Change Point Inference and Identification for High-Dimensional Mean Vectors. Journal of the Royal Statistical Society Series B: Statistical Methodology, 83, 247–270.
}
\seealso{
\code{\link{cal_YC21}}
}
