% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{ZWS22}
\alias{ZWS22}
\title{ZWS22}
\usage{
ZWS22(X)
}
\arguments{
\item{X}{\eqn{n\times p}}
}
\value{
The p-value
}
\description{
ZWS22
}
\references{
Zhang, Y., Wang, R. and Shao, X. (2022) Adaptive Inference for Change Points in High-Dimensional Data. Journal of the American Statistical Association, 117, 1751–1762.
}
\seealso{
\code{\link{cal_ZWS22}}
}
